IntervalFireStats: compute stats instead of hashes (#14)
The previous implementation in IntervalFireSHA is useful when results
are expected to be bitwise identical between NEURON and CoreNEURON. In
the presence of some compiler optimisations this is not a realistic
standard to expect. IntervalFireStats instead computes the mean time in
between events, and the test is modified to allow *small* differences in
this quantity between different runs.